+++ db.go
@@ -347,115 +347,3 @@ func quoteIdentifier(name string, driver string) string {
 
 	return strings.Join(result, ".")
 }
-
-// PrepareInsertStatement is a prepared insert statement for entity
-type PrepareInsertStatement struct {
-	md       *Metadata
-	stmt     *sqlx.NamedStmt
-	dbDriver string
-}
-
-// Close closes the prepared statement
-func (pis *PrepareInsertStatement) Close() error {
-	return pis.stmt.Close()
-}
-
-// ExecContext executes a prepared insert statement using the Entity passed.
-func (pis *PrepareInsertStatement) ExecContext(ctx context.Context, ent Entity) (lastID int64, err error) {
-	ctx, cancel := context.WithTimeout(ctx, WriteTimeout)
-	defer cancel()
-
-	if err := ent.OnEntityEvent(ctx, EventBeforeInsert); err != nil {
-		return 0, fmt.Errorf("before insert, %w", err)
-	}
-
-	lastID, err = pis.execContext(ctx, ent)
-	if err != nil {
-		if isConflictError(err, pis.dbDriver) {
-			return 0, ErrConflict
-		}
-		return 0, err
-	}
-
-	if err := ent.OnEntityEvent(ctx, EventAfterInsert); err != nil {
-		return 0, fmt.Errorf("after insert, %w", err)
-	}
-
-	return lastID, nil
-}
-
-func (pis *PrepareInsertStatement) execContext(ctx context.Context, ent Entity) (lastID int64, err error) {
-	if pis.md.hasReturningInsert {
-		err := pis.stmt.QueryRowxContext(ctx, ent).StructScan(ent)
-		return 0, err
-	}
-
-	result, err := pis.stmt.ExecContext(ctx, ent)
-	if err != nil {
-		return 0, err
-	} else if pis.dbDriver == driverPostgres {
-		// postgresql不支持LastInsertId特性
-		return 0, nil
-	}
-
-	lastID, err = result.LastInsertId()
-	if err != nil {
-		return 0, fmt.Errorf("get last insert id, %w", err)
-	}
-	return lastID, nil
-}
-
-// PrepareUpdateStatement is a prepared update statement for entity
-type PrepareUpdateStatement struct {
-	md       *Metadata
-	stmt     *sqlx.NamedStmt
-	dbDriver string
-}
-
-// Close closes the prepared statement
-func (pus *PrepareUpdateStatement) Close() error {
-	return pus.stmt.Close()
-}
-
-// ExecContext executes a prepared update statement using the Entity passed.
-func (pus *PrepareUpdateStatement) ExecContext(ctx context.Context, ent Entity) error {
-	ctx, cancel := context.WithTimeout(ctx, WriteTimeout)
-	defer cancel()
-
-	if err := ent.OnEntityEvent(ctx, EventBeforeUpdate); err != nil {
-		return fmt.Errorf("before update, %w", err)
-	}
-
-	if err := pus.execContext(ctx, ent); err != nil {
-		return err
-	}
-
-	if v, ok := ent.(Cacheable); ok {
-		if err := DeleteCache(v); err != nil {
-			return fmt.Errorf("delete cache, %w", err)
-		}
-	}
-
-	if err := ent.OnEntityEvent(ctx, EventAfterUpdate); err != nil {
-		return fmt.Errorf("after update, %w", err)
-	}
-	return nil
-}
-
-func (pus *PrepareUpdateStatement) execContext(ctx context.Context, ent Entity) error {
-	if pus.md.hasReturningUpdate {
-		return pus.stmt.QueryRowxContext(ctx, ent).StructScan(ent)
-	}
-
-	result, err := pus.stmt.ExecContext(ctx, ent)
-	if err != nil {
-		return err
-	}
-
-	if n, err := result.RowsAffected(); err != nil {
-		return fmt.Errorf("get affected rows, %w", err)
-	} else if n == 0 {
-		return sql.ErrNoRows
-	}
-	return nil
-}
